+++ lib/__prototype.js
@@ -8,6 +8,7 @@ var fetch = require('./fetch');
 var types = require('./types');
 var joinUrl = require('proper-url-join');
 var middlewares = require('./middlewares');
+var builders = require('./builders');
 
 function set(key, value) {
   if (types.any.includes(key)) {
@@ -91,6 +92,11 @@ function use(middleware) {
   return this;
 }
 
+function build(builder) {
+  if (typeof builder == 'function') builder(this);else if (builders[builder]) builders[builder](this);
+  return this;
+}
+
 var _ = {
   url: url,
   set: set,
@@ -98,6 +104,7 @@ var _ = {
   add: add,
   push: push,
   use: use,
+  build: build,
   alias: alias,
   key: key,
