+++ lib/builders.js
@@ -1,13 +1,10 @@
 'use strict';
 
-Object.defineProperty(exports, "__esModule", {
-  value: true
-});
-exports.default = {
+module.exports = {
   add: function add(name, builder) {
 
     // @TODO: invalid params
-    if (typeof name == 'string' || typeof builder !== 'function') return;
+    if (typeof name != 'string' || typeof builder !== 'function') return;
 
     // @TODO: name exist
     //   this[name]
