+++ test/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html><html>
+<head>
+<script  type="text/javascript" src="../dist/es5.js" ></script>
+
+
+</head>
+<body>
+</body>
+</html>
